checkout: modify lookup by package
The test CheckoutScanCase.test_scan_document_error_different_picking_type_package
was no longer passing because was returning a "barcode not
found" instead of "you cannot move this using this menu".

The picking type should not be checked at this point, as it is checked
in _select_picking() (called at the end of the method), which will
return the appropriate message.

Note: replaced the search by a field to reuse the cache if we had such
data in cache.