shopfloor: zone picking add unloading on first screen
The user may end up in the `scan_location` screen with some products to
unload. It can happen when clicking on the back button but also when
there is no lines left to process in the current zone.

This change adds an Unload At Destination button on the `scan_location`
screen which will be display only when the user has some product to
unload.